refactor: Migrate Invoice model, manager, serializer, viewsets to credit_cards app
