+++ poupeai_finance_service/credit_cards/api/viewsets.py
@@ -1,10 +1,15 @@
+from rest_framework import mixins, viewsets
 from rest_framework.viewsets import ModelViewSet
 from rest_framework.permissions import IsAuthenticated
+from rest_framework import filters
+from django_filters.rest_framework import DjangoFilterBackend
 
-from poupeai_finance_service.credit_cards.models import CreditCard
+from poupeai_finance_service.credit_cards.models import CreditCard, Invoice
 from poupeai_finance_service.credit_cards.api.serializers import CreditCardSerializer
 from poupeai_finance_service.users.api.permissions import IsProfileActive
+from poupeai_finance_service.core.permissions import IsOwnerProfile
 from poupeai_finance_service.users.querysets import get_profile_by_user
+from poupeai_finance_service.credit_cards.api.serializers import InvoiceSerializer
 
 class CreditCardViewSet(ModelViewSet):
     queryset = CreditCard.objects.all()
@@ -23,3 +28,24 @@ class CreditCardViewSet(ModelViewSet):
         if self.request.user.is_authenticated:
             context['profile'] = get_profile_by_user(self.request.user)
         return context
+
+class InvoiceViewSet(mixins.RetrieveModelMixin,
+                     mixins.ListModelMixin,
+                     viewsets.GenericViewSet):
+    """
+    Viewset for retrieving and listing Invoices.
+    Invoices are primarily generated by the system for credit card transactions.
+    """
+    queryset = Invoice.objects.all()
+    serializer_class = InvoiceSerializer
+    permission_classes = [IsAuthenticated, IsOwnerProfile]
+    filter_backends = [DjangoFilterBackend, filters.OrderingFilter]
+    ordering_fields = ['due_date', 'month', 'year', 'amount_paid']
+    ordering = ['-year', '-month']
+
+    def get_queryset(self):
+        """
+        Ensures a user can only see invoices related to their credit cards.
+        """
+        user_profile = self.request.user.profile
+        return self.queryset.filter(credit_card__profile=user_profile)
