+++ poupeai_finance_service/credit_cards/models.py
@@ -5,6 +5,7 @@ from django.utils.translation import gettext_lazy as _
 from poupeai_finance_service.core.models import TimeStampedModel
 from poupeai_finance_service.users.models import Profile
 from poupeai_finance_service.credit_cards.validators import validate_day, validate_closing_due_days_not_equal
+from .managers import InvoiceManager
 
 class CreditCard(TimeStampedModel):
     class BrandChoices(models.TextChoices):
@@ -72,3 +73,41 @@ class CreditCard(TimeStampedModel):
             validate_closing_due_days_not_equal(self.closing_day, self.due_day)
         except ValidationError as e:
             raise ValidationError({'due_day': e.message})
+
+class Invoice(TimeStampedModel):
+    """
+    Model representing a credit card invoice.
+    """
+    credit_card = models.ForeignKey(
+        CreditCard,
+        on_delete=models.CASCADE,
+        related_name='invoices',
+        verbose_name=_('Credit Card')
+    )
+    month = models.SmallIntegerField(_('Month'), validators=[MinValueValidator(1), MinValueValidator(12)])
+    year = models.SmallIntegerField(_('Year'), validators=[MinValueValidator(2000)])
+    amount_paid = models.DecimalField(
+        _('Amount Paid'),
+        max_digits=10,
+        decimal_places=2,
+        default=0,
+        validators=[MinValueValidator(0)]
+    )
+    due_date = models.DateField(_('Due Date'))
+    paid = models.BooleanField(_('Paid'), default=False)
+
+    objects = InvoiceManager()
+
+    class Meta:
+        verbose_name = _('Invoice')
+        verbose_name_plural = _('Invoices')
+        unique_together = ('credit_card', 'month', 'year')
+        ordering = ['-year', '-month']
+
+    def __str__(self):
+        return f"{self.credit_card.name} - {self.month}/{self.year}"
+
+    @property
+    def total_amount(self):
+        """Calculates the total amount of the invoice based on associated credit card transactions."""
+        return self.transactions.aggregate(total=models.Sum('amount'))['total'] or 0
