+++ poupeai_finance_service/transactions/api/viewsets.py
@@ -13,12 +13,11 @@ from django.utils import timezone
 from django.core.exceptions import ValidationError
 
 from poupeai_finance_service.transactions.services import TransactionService
-from poupeai_finance_service.transactions.models import Transaction, Invoice
+from poupeai_finance_service.transactions.models import Transaction
 from poupeai_finance_service.transactions.api.serializers import (
     TransactionListSerializer,
     TransactionDetailSerializer,
-    TransactionCreateUpdateSerializer,
-    InvoiceSerializer
+    TransactionCreateUpdateSerializer
 )
 from poupeai_finance_service.core.permissions import IsOwnerProfile
 
@@ -103,27 +102,6 @@ class TransactionViewSet(viewsets.ModelViewSet):
             raise DRFValidationError(e.detail)
         except Exception as e:
             return Response(
-                {"detail": _(f"Ocorreu um erro ao tentar deletar a transação: {e}")},
+                {"detail": _(f"An error occurred while trying to delete the transaction: {e}")},
                 status=status.HTTP_500_INTERNAL_SERVER_ERROR
             )
-
-class InvoiceViewSet(mixins.RetrieveModelMixin,
-                     mixins.ListModelMixin,
-                     viewsets.GenericViewSet):
-    """
-    Viewset for retrieving and listing Invoices.
-    Invoices are primarily generated by the system for credit card transactions.
-    """
-    queryset = Invoice.objects.all()
-    serializer_class = InvoiceSerializer
-    permission_classes = [IsAuthenticated, IsOwnerProfile]
-    filter_backends = [DjangoFilterBackend, filters.OrderingFilter]
-    ordering_fields = ['due_date', 'month', 'year', 'amount_paid']
-    ordering = ['-year', '-month']
-
-    def get_queryset(self):
-        """
-        Ensures a user can only see invoices related to their credit cards.
-        """
-        user_profile = self.request.user.profile
-        return self.queryset.filter(credit_card__profile=user_profile)
