+++ poupeai_finance_service/transactions/urls.py
@@ -1,12 +1,11 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from poupeai_finance_service.transactions.api.viewsets import TransactionViewSet, InvoiceViewSet
+from poupeai_finance_service.transactions.api.viewsets import TransactionViewSet
 
 app_name = 'transactions'
 
 router = DefaultRouter()
 router.register(r'', TransactionViewSet, basename='transactions')
-router.register(r'invoices', InvoiceViewSet, basename='invoices')
 
 app_name = 'transactions'
